+++ File.swift
@@ -1,8 +1,0 @@
-//
-//  File.swift
-//  Puddles
-//
-//  Created by Dennis Müller on 23.03.23.
-//
-
-import Foundation