+++ Sources/Puddles/Queryables/QueryError.swift
@@ -2,7 +2,3 @@ import Foundation
 
 /// An error indicating the cancellation of a query.
 public struct QueryCancellationError: Swift.Error {}
-
-enum QueryInternalError: Swift.Error {
-    case queryAutoCancel
-}